+++ server/src/database/run-seeder.ts
@@ -1,8 +1,6 @@
 import { AppModule } from 'src/app.module';
 import { AccountAuth } from 'src/modules/accounts/entities/account-auth.entity';
 import { Account } from 'src/modules/accounts/entities/account.entity';
-import { Portfolio } from 'src/modules/portfolios/enities/portfolio.entity';
-import { TradeRecord } from 'src/modules/portfolios/enities/trade-record.entity';
 import { Role } from 'src/modules/privileges/entities/role.entity';
 import {
     DataSource,
@@ -15,6 +13,11 @@ import {
 
 import { ConfigService } from '@nestjs/config';
 import { NestFactory } from '@nestjs/core';
+import { Portfolio } from 'src/modules/portfolios/enities/portfolio.entity';
+import {
+    CashTradeRecord,
+    FXTradeRecord, StockTradeRecord, 
+} from 'src/modules/portfolios/enities/trade-record.entity';
 
 async function run() {
     const app = await NestFactory.createApplicationContext(AppModule);
@@ -32,7 +35,9 @@ async function run() {
             AccountAuth,
             Role,
             Portfolio,
-            TradeRecord,
+            StockTradeRecord,
+            FXTradeRecord,
+            CashTradeRecord,
         ],
 
         seeds: ['src/database/seeds/**/*{.ts,.js}'],