+++ server/src/enums/transaction.enum.ts
@@ -3,7 +3,15 @@ export enum EnumTradeDirection {
     SELL = 'sell',
 };
 
+export enum EnumCashFlow {
+    /** 存入 */
+    DEPOSIT = 'deposit',
+    /** 取出 */
+    WITHDRAW = 'withdraw'
+}
+
 export enum EnumAssetType {
     CASH = 'cash',
+    FX = 'fx',
     STOCK = 'stock'
 }