+++ server/src/modules/portfolios/dto/create-portfolio.dto.ts
@@ -1,4 +1,5 @@
 import {
+    IsDateString,
     IsNotEmpty,
     IsNumber,
     IsString,
@@ -16,9 +17,18 @@ export class CreatePortfolioDto {
     @ApiProperty({ description: 'name of the portfolio' })
         title: string;
 
+    @IsDateString()
+    @IsNotEmpty()
+    @ApiProperty({ description: 'init date of the portfolio' })
+        createDate: string;
+            
+    @IsString()
+    @IsNotEmpty()
+    @ApiProperty({ description: 'base currency of the portfolio' })
+        currency: string;
+            
     @IsNumber()
     @IsNotEmpty()
     @ApiProperty({ description: 'Initial balance of portfolio' })
         initialBalance: number;
-
 }