+++ server/src/modules/portfolios/dto/update-portfolio.dto.ts
@@ -1,0 +1,15 @@
+import { IsString } from 'class-validator';
+
+import {
+    ApiProperty,
+    ApiSchema,
+} from '@nestjs/swagger';
+
+@ApiSchema({ name: 'UpdatePortfolioRequest' })
+export class UpdatePortfolioDto {
+    @IsString()
+    @ApiProperty({
+        required: false, description: 'name of the portfolio', 
+    })
+        title?: string;
+}