+++ server/src/modules/portfolios/enities/trade-record.entity.ts
@@ -1,6 +1,7 @@
 import {
     EnumAssetType,
     EnumTradeDirection,
+    EnumCashFlow,
 } from 'src/enums';
 import {
     Column,
@@ -19,13 +20,17 @@ import {
 
 import { Portfolio } from './portfolio.entity';
 
-@ApiSchema({ name: 'TradeRecordDto' })
+@ApiSchema({ name: 'CashTradeRecordDto' })
 @Entity()
-export class TradeRecord {
-    constructor(auth: Partial<TradeRecord>) {
-        Object.assign(this, auth);
+export class CashTradeRecord {
+    constructor(data: Partial<CashTradeRecord>) {
+        Object.assign(this, data);
     }
 
+    @ApiProperty()
+    @PrimaryGeneratedColumn()
+        id: number;
+        
     @ApiProperty()
     @CreateDateColumn()
         create_at: Date;
@@ -36,42 +41,75 @@ export class TradeRecord {
     })
         update_at: Date;
 
+    @Column({ type: 'bigint' })
     @ApiProperty()
-    @PrimaryGeneratedColumn()
-        id: number;
+        portfolio_id: number;
 
-    @ApiProperty({ description: '交易類型' })
+    @ApiProperty({ description: '資產類型' })
     @Column({ enum: EnumAssetType })
         asset_type: EnumAssetType;
 
-    @ApiProperty({ description: '股票代號' })
+    @ApiProperty({
+        description: '存入/領回', enum: EnumCashFlow, 
+    })
+    @Column({ enum: EnumCashFlow })
+        direction: EnumCashFlow;
+            
+    @ApiProperty({ description: '手續費' })
     @Column()
-        symbol: string;
-
-    @Column({ type: 'bigint' })
-        portfolio_id: number;
+        commission: number;
 
     @ApiProperty({ description: '交易日期' })
-    @Column()
-        trade_date: Date;
-    
-    @ApiProperty({
-        description: '交易方向', enum: EnumTradeDirection, 
-    })
-    @Column({ enum: EnumTradeDirection })
-        direction: string;
+    @Column({ type: 'date' })
+        trade_date: string;
 
-    @ApiProperty({ description: '交易價位' })
-    @Column()
-        execution_price: number;
-
-    @ApiProperty({ description: '交易數量(股)' })
+    @ApiProperty({ description: '交易數量' })
     @Column()
         quantity: number;
 
-    @ApiProperty({ description: '交易總價 = 交易價位 * 交易數量' })
-    @Column()
-        total_amount: number;
+    @ManyToOne(() => Portfolio, (portfolio) => portfolio.stockTradeRecords)
+    @JoinColumn({ name: 'portfolio_id' })
+        portfolio: Portfolio;
+}
+
+@ApiSchema({ name: 'StockTradeRecordDto' })
+@Entity()
+export class StockTradeRecord {
+    constructor(data: Partial<StockTradeRecord>) {
+        Object.assign(this, data);
+    }
+
+    @ApiProperty()
+    @PrimaryGeneratedColumn()
+        id: number;
+        
+    @ApiProperty()
+    @CreateDateColumn()
+        create_at: Date;
+
+    @ApiProperty()
+    @UpdateDateColumn({
+        nullable: true, default: null, 
+    })
+        update_at: Date;
+
+    @Column({ type: 'bigint' })
+    @ApiProperty()
+        portfolio_id: number;
+
+    @ApiProperty({ description: '資產類型' })
+    @Column({ enum: EnumAssetType })
+        asset_type: EnumAssetType;
+            
+    @ApiProperty({
+        description: '買進/賣出', enum: EnumTradeDirection, 
+    })
+    @Column({ enum: EnumTradeDirection })
+        direction: EnumTradeDirection;
+        
+    @ApiProperty({ description: '交易日期' })
+    @Column({ type: 'date' })
+        trade_date: string;
 
     @ApiProperty({ description: '手續費' })
     @Column()
@@ -81,8 +119,107 @@ export class TradeRecord {
     @Column()
         tax: number;
 
-    @ManyToOne(() => Portfolio, (portfolio) => portfolio.tradeRecords)
+    @ApiProperty({ description: '股票代號' })
+    @Column()
+        symbol: string;
+
+    @ApiProperty({ description: '交易價位' })
+    @Column()
+        execution_price: number;
+
+    @ApiProperty({ description: '交易數量(股)' })
+    @Column()
+        quantity: number;
+
+    @ApiProperty({ description: '成本' })
+    @Column()
+        cost: number;
+
+    @ApiProperty({ description: '已實現損益' })
+    @Column()
+        realized_profit_loss: number;
+
+    @ManyToOne(() => Portfolio, (portfolio) => portfolio.stockTradeRecords)
     @JoinColumn({ name: 'portfolio_id' })
         portfolio: Portfolio;
 }
 
+@ApiSchema({ name: 'FXTradeRecordDto' })
+@Entity()
+export class FXTradeRecord {
+    constructor(data: Partial<FXTradeRecord>) {
+        Object.assign(this, data);
+    }
+
+    @ApiProperty()
+    @PrimaryGeneratedColumn()
+        id: number;
+
+    @ApiProperty()
+    @CreateDateColumn()
+        create_at: Date;
+
+    @ApiProperty()
+    @UpdateDateColumn({
+        nullable: true, default: null, 
+    })
+        update_at: Date;
+        
+    @Column({ type: 'bigint' })
+    @ApiProperty()
+        portfolio_id: number;
+
+    @ApiProperty({ description: '資產類型' })
+    @Column({ enum: EnumAssetType })
+        asset_type: EnumAssetType;
+            
+    @ApiProperty({
+        description: '買進/賣出', enum: EnumTradeDirection, 
+    })
+    @Column({ enum: EnumTradeDirection })
+        direction: EnumTradeDirection;
+        
+    @ApiProperty({ description: '交易日期' })
+    @Column({ type: 'date' })
+        trade_date: string;
+
+    @ApiProperty({ description: '手續費' })
+    @Column()
+        commission: number;
+    
+    @ApiProperty({ description: '稅' })
+    @Column()
+        tax: number;
+
+    @ApiProperty({ description: '本金幣別' })
+    @Column()
+        base_currency: string;
+
+    @ApiProperty({ description: '交易幣別' })
+    @Column()
+        target_currency: string;
+
+    @ApiProperty({ description: '匯率' })
+    @Column()
+        exchange_rate: number;
+
+    @ApiProperty({ description: '本金幣別數量' })
+    @Column()
+        base_quantity: number;
+
+    @ApiProperty({ description: '目標幣別數量' })
+    @Column()
+        target_quantity: number;
+
+    @ApiProperty({ description: '成本(本金幣別)' })
+    @Column()
+        cost: number;
+
+    @ApiProperty({ description: '已實現損益' })
+    @Column()
+        realized_profit_loss: number;
+
+    @ManyToOne(() => Portfolio, (portfolio) => portfolio.fxTradeRecords)
+    @JoinColumn({ name: 'portfolio_id' })
+        portfolio: Portfolio;
+}